requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
jinja2
python-dotenv
aiocron